Add files via upload
This is my work on the assignment of course 5 in the IBM Data Analyst Certification 9-course series. In this notebook, I used web scraping with Beautiful Soup and the requests library to obtain the revenue data of Tesla and GameStop. I also used the yfinance library to get the stock price data of both companies, and then I visualized them.